+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Removed
 
+* Registration module, Editing (manual): remove option to plot registration matrix.
+
 ### Fixed
 
 * Registration module with "timepoint range" option: mark timepoints outside of the selected range as ignored in the transformation matrix (set column align_0_1 to 0). For the registered image, save only only time frames within the selected range.